[ntcore] Split LocalStorage implementation into separate files
